requirements.txt
python-jobspy
pandas
tqdm